Update lints with better rules
