+++ tslint.json
@@ -13,9 +13,8 @@
 	"rules": {
 		// Try to keep synced with Flutter extension!
 		"disallow-fspath": true,
-		"disallow-extension-code-in-tests": true,
-		"disallow-vscode-in-debuggers": true,
-		"disallow-vscode-in-utils": true,
+		"disallow-importing-non-shared-code": true,
+		"disallow-vscode": true,
 		"indent": {
 			"options": "tabs"
 		},
